requirements/requirements_doc.txt
#PyMechanical
ansys-mechanical-core[doc, viz]==0.11.1

# Documentation dependencies
pyvista[jupyter]==0.43.10
ipyvtklink==0.2.3
